Remove the code that detects modules added after initial load.
This is too much code complexity for too little gain (covering the case where a simulator temporarily has all its regions removed but is not shutdown, it's not a major requirement).
Removing this check also at least allows the module to work with dynamically added modules without restarting (as long as modules never falls to zero).